feat: only allow users in a related establishment to access dashboard

+++ src/modules/business-intelligence/dashboard/dashboard.controller.ts
@@ -1,4 +1,7 @@
 import { PaginationQuery } from '@libs/types/pagination'
+import { AuthUser } from '@modules/identity/auth/decorators/auth-user.decorator'
+import { JwtAuthGuard } from '@modules/identity/auth/guards/jwt-auth.guard'
+import { User } from '@modules/identity/user/entities/user.entity'
 import {
   Controller,
   Get,
@@ -9,17 +12,33 @@ import {
   Delete,
   ParseIntPipe,
   Query,
-  NotFoundException
+  NotFoundException,
+  UseGuards,
+  ForbiddenException
 } from '@nestjs/common'
-import { ApiTags } from '@nestjs/swagger'
+import { ApiBearerAuth, ApiTags } from '@nestjs/swagger'
 import { DashboardService } from './dashboard.service'
 import { CreateDashboardDto } from './dto/create-dashboard.dto'
 import { UpdateDashboardDto } from './dto/update-dashboard.dto'
+import { Dashboard } from './entities/dashboard.entity'
 
 @Controller('dashboards')
 @ApiTags('Dashboard')
+@ApiBearerAuth()
+@UseGuards(JwtAuthGuard)
 export class DashboardController {
   constructor(private readonly dashboardService: DashboardService) {}
+
+  async canSee(user: User, dashboard: Dashboard) {
+    const userInRelatedEstablishment =
+      (
+        await dashboard.establishmentsWithAccess.matching({
+          where: { workRelations: { worker: user.medicalWorker } }
+        })
+      ).length > 0
+
+    return user.isAdmin || userInRelatedEstablishment
+  }
 
   @Post()
   create(@Body() createDashboardDto: CreateDashboardDto) {
@@ -27,22 +46,56 @@ export class DashboardController {
   }
 
   @Get()
-  findAll(@Query() query: PaginationQuery) {
-    return this.dashboardService.findAll(query)
+  findAll(@Query() query: PaginationQuery, @AuthUser() currentUser: User) {
+    return this.dashboardService.findAll(query, currentUser)
   }
 
   @Get(':id')
-  findOne(@Param('id', ParseIntPipe) id: number) {
-    return this.dashboardService.findOne(id)
+  async findOne(@Param('id') id: number, @AuthUser() currentUser: User) {
+    const dashboard = await this.dashboardService.findOne(id)
+
+    if (!dashboard) {
+      throw new NotFoundException()
+    }
+
+    if (!(await this.canSee(currentUser, dashboard))) {
+      throw new ForbiddenException()
+    }
+
+    return dashboard
   }
 
   @Patch(':id')
-  update(@Param('id', ParseIntPipe) id: number, @Body() updateDashboardDto: UpdateDashboardDto) {
+  async update(
+    @Param('id', ParseIntPipe) id: number,
+    @Body() updateDashboardDto: UpdateDashboardDto,
+    @AuthUser() currentUser: User
+  ) {
+    const dashboard = await this.dashboardService.findOne(id)
+
+    if (!dashboard) {
+      throw new NotFoundException()
+    }
+
+    if (!(await this.canSee(currentUser, dashboard))) {
+      throw new ForbiddenException()
+    }
+
     return this.dashboardService.update(id, updateDashboardDto)
   }
 
   @Delete(':id')
-  remove(@Param('id', ParseIntPipe) id: number) {
+  async remove(@Param('id', ParseIntPipe) id: number, @AuthUser() currentUser: User) {
+    const dashboard = await this.dashboardService.findOne(id)
+
+    if (!dashboard) {
+      throw new NotFoundException()
+    }
+
+    if (!(await this.canSee(currentUser, dashboard))) {
+      throw new ForbiddenException()
+    }
+
     const couldRemove = this.dashboardService.remove(id)
     if (!couldRemove) {
       throw new NotFoundException()
